+++ src/Patch/DefinitionList/Loader/ComponentPool.php
@@ -71,6 +71,7 @@ class ComponentPool
             'bundle' => new LoaderComponents\BundleComponent($rootPackage),
             'global-exclude' => $excludes ? new LoaderComponents\GlobalExcludeComponent($excludes) : false,
             'local-exclude' => new LoaderComponents\LocalExcludeComponent(),
+            'root-patch' => new LoaderComponents\RootPatchComponent($rootPackage),
             'path-normalizer' => new LoaderComponents\PathNormalizerComponent($packageInfoResolver),
             'constraints' => new LoaderComponents\ConstraintsComponent($configExtractor),
             'downloader' => new LoaderComponents\DownloaderComponent($rootPackage, $downloader),